requirements.txt
# 核心依赖
torch>=2.0.0
transformers>=4.37.0
accelerate>=0.25.0
peft>=0.7.0
bitsandbytes>=0.41.0

# 数据处理
datasets>=2.14.0
pandas>=1.5.0
numpy>=1.24.0

# 评估工具
scikit-learn>=1.3.0
tqdm>=4.65.0

# 工具库
pyyaml>=6.0
sentencepiece>=0.1.99

# 可选：用于数据可视化
matplotlib>=3.7.0
seaborn>=0.12.0

